hw2/requirements.txt
omegaconf>=2.1.0
fvcore>=0.1.5
iopath>=0.1.9
hydra-core>=1.1.0
pycocotools>=2.0.6
pyyaml>=5.4
Pillow>=9.0.0
opencv-python>=4.6.0
scipy>=1.7.0
einops>=0.6.0
timm>=0.9.0
fairscale>=0.4.0
pybind11>=2.10.0
tensorboard>=2.11.0
tqdm>=4.64.0
matplotlib>=3.5.0
albumentations>=1.3.0
pandas>=1.5.0
setuptools<82
